% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{coco_method}
\alias{coco_method}
\title{Choose Optimization Method}
\usage{
coco_method(
  model,
  method = c("standard", "differential_evolution", "particle_swarm", "bee_colony",
    "genetic_algorithm", "jaya", "cma_es", "generalized_annealing", "bayesian")
)
}
\arguments{
\item{model}{A coconut_spec object.}

\item{method}{Optimization algorithm: "standard", "differential_evolution", "particle_swarm", "bee_colony", "genetic_algorithm", "jaya", "cma_es", "generalized_annealing", or "bayesian".}
}
\value{
Updated coconut_spec object.
}
\description{
Choose Optimization Method
}
