% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{coco_mode}
\alias{coco_mode}
\title{Create a Coconut Optimization Mode}
\usage{
coco_mode(objective, mode = c("minimize", "maximize"))
}
\arguments{
\item{objective}{Function to optimize (minimize or maximize).}

\item{mode}{Optimization mode: "minimize" or "maximize".}
}
\value{
A coconut_spec object.
}
\description{
Create a Coconut Optimization Mode
}
