% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{coco_params}
\alias{coco_params}
\title{Set Optimization Parameters}
\usage{
coco_params(model, ...)
}
\arguments{
\item{model}{A coconut_spec object.}

\item{...}{Standardized parameters (lower, upper, pop_size, max_iter, initial) or method-specific controls (via method_control).}
}
\value{
Updated coconut_spec object.
}
\description{
Set Optimization Parameters
}
