\name{CoCoNuT-package}
\alias{CoCoNuT-package}
\alias{CoCoNuT}
\docType{package}
\title{
\packageTitle{CoCoNuT}
}
\description{
\packageDescription{CoCoNuT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CoCoNuT}
\packageIndices{CoCoNuT}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{CoCoNuT}

Maintainer: \packageMaintainer{CoCoNuT}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
