% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{rep_data_gen}
\alias{rep_data_gen}
\title{Simulate data from normal distribution for replicability analysis: one auxiliary covariate}
\usage{
rep_data_gen(
  m = 10000,
  xi = c(0.9, 0.025, 0.025, 0.05),
  info.str = "moderate",
  mu1 = 2,
  mu2 = 2,
  mu3 = 2
)
}
\arguments{
\item{m}{the number of features to be simulated.}

\item{xi}{a numeric vector of the prior probabilities for the hidden states of two primary studies corresponding to (0, 0), (0, 1), (1, 0), (1, 1).}

\item{info.str}{a string representing the informative strength of the auxiliary covariate: "uninormative", "weak", "moderate", "strong".}

\item{mu1}{a numeric value of the signal strength in study 1.}

\item{mu2}{a numeric value of the signal strength in study 2.}

\item{mu3}{a numeric value of the signal strength in the auxiliary covariate.}
}
\value{
a list with the following elements:
\item{pvals1}{a numeric vector of the simulated p-values for study 1.}
\item{pvals2}{a numeric vector of the simulated p-values for study 2.}
\item{x}{a numeric vector of the simulated p-values for the auxiliary study.}
\item{theta1}{a numeric vector of 0/1 indicating the hidden states of study 1.}
\item{theta2}{a numeric vector of 0/1 indicating the hidden states of study 2.}
\item{zeta}{a numeric vector of 0/1 indicating the hidden states of the auxiliary study.}
}
\description{
Simulate data from normal distribution for replicability analysis: one auxiliary covariate
}
