% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{rep_data_gen_multi}
\alias{rep_data_gen_multi}
\title{Simulate data from normal distribution for replicability analysis: two auxiliary covariates}
\usage{
rep_data_gen_multi(
  m = 10000,
  xi = c(0.9, 0.025, 0.025, 0.05),
  info.str = c("moderate", "moderate"),
  mu1 = 2,
  mu2 = 2,
  mu3 = 2,
  mu4 = 2
)
}
\arguments{
\item{m}{the number of features to be simulated.}

\item{xi}{a numeric vector of the prior probabilities for the hidden states of two primary studies corresponding to (0, 0), (0, 1), (1, 0), (1, 1).}

\item{info.str}{a string vector representing the informative strengths of two auxiliary covariates, with potential elements: "uninormative", "weak", "moderate", "strong".}

\item{mu1}{a numeric value of the signal strength in primary study 1.}

\item{mu2}{a numeric value of the signal strength in primary study 2.}

\item{mu3}{a numeric value of the signal strength in the auxiliary study 1.}

\item{mu4}{a numeric value of the signal strength in the auxiliary study 2.}
}
\value{
a list with the following elements:
\item{pvals1}{a numeric vector of the simulated p-values for primary study 1.}
\item{pvals2}{a numeric vector of the simulated p-values for primary study 2.}
\item{x1}{a numeric vector of the simulated p-values for auxiliary study 1.}
\item{x2}{a numeric vector of the simulated p-values for auxiliary study 2.}
\item{theta1}{a numeric vector of 0/1 indicating the hidden states of primary study 1.}
\item{theta2}{a numeric vector of 0/1 indicating the hidden states of primary study 2.}
\item{zeta1}{a numeric vector of 0/1 indicating the hidden states of auxiliary study 1.}
\item{zeta2}{a numeric vector of 0/1 indicating the hidden states of auxiliary study 2.}
}
\description{
Simulate data from normal distribution for replicability analysis: two auxiliary covariates
}
