% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coconut-package.R
\docType{package}
\name{coconut-package}
\alias{coconut}
\alias{coconut-package}
\title{coconut: Collection Of Complex Optimizers as a Nice Unified Toolkit}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

coconut is designed as unified optimization framework. inspired by other pipeable interfaces in R, but for optimizers.
}
\author{
\strong{Maintainer}: frank hull \email{frankiethull@gmail.com}

}
\keyword{internal}
